+++ src/main/java/cloud/metaapi/sdk/meta_api/MetatraderAccountApi.java
@@ -5,7 +5,6 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.stream.Collectors;
 
-import cloud.metaapi.sdk.clients.meta_api.ExpertAdvisorClient;
 import cloud.metaapi.sdk.clients.meta_api.MetaApiWebsocketClient;
 import cloud.metaapi.sdk.clients.meta_api.MetatraderAccountClient;
 import cloud.metaapi.sdk.clients.meta_api.models.AccountsFilter;
@@ -19,7 +18,6 @@ public class MetatraderAccountApi {
   private MetatraderAccountClient metatraderAccountClient;
   private MetaApiWebsocketClient metaApiWebsocketClient;
   private ConnectionRegistry connectionRegistry;
-  private ExpertAdvisorClient expertAdvisorClient;
   
   /**
    * Constructs a MetaTrader account API instance
@@ -29,12 +27,10 @@ public class MetatraderAccountApi {
    * @param expertAdvisorClient expert advisor REST API client
    */
   public MetatraderAccountApi(MetatraderAccountClient metatraderAccountClient,
-    MetaApiWebsocketClient metaApiWebsocketClient, ConnectionRegistry connectionRegistry,
-    ExpertAdvisorClient expertAdvisorClient) {
+    MetaApiWebsocketClient metaApiWebsocketClient, ConnectionRegistry connectionRegistry) {
     this.metatraderAccountClient = metatraderAccountClient;
     this.metaApiWebsocketClient = metaApiWebsocketClient;
     this.connectionRegistry = connectionRegistry;
-    this.expertAdvisorClient = expertAdvisorClient;
   }
   
   /**
@@ -53,7 +49,7 @@ public class MetatraderAccountApi {
   public CompletableFuture<List<MetatraderAccount>> getAccounts(AccountsFilter accountsFilter) {
     return metatraderAccountClient.getAccounts(accountsFilter).thenApply(accounts -> {
       return accounts.stream().map(accountDto -> new MetatraderAccount(accountDto, metatraderAccountClient,
-        metaApiWebsocketClient, connectionRegistry, expertAdvisorClient)).collect(Collectors.toList());
+        metaApiWebsocketClient, connectionRegistry)).collect(Collectors.toList());
     });
   }
   
@@ -65,7 +61,7 @@ public class MetatraderAccountApi {
   public CompletableFuture<MetatraderAccount> getAccount(String accountId) {
     return metatraderAccountClient.getAccount(accountId).thenApply(accountDto -> {
       return new MetatraderAccount(accountDto, metatraderAccountClient, metaApiWebsocketClient,
-        connectionRegistry, expertAdvisorClient);
+        connectionRegistry);
     });
   }
   
@@ -76,7 +72,7 @@ public class MetatraderAccountApi {
   public CompletableFuture<MetatraderAccount> getAccountByToken() {
     return metatraderAccountClient.getAccountByToken().thenApply(accountDto -> {
       return new MetatraderAccount(accountDto, metatraderAccountClient, metaApiWebsocketClient,
-        connectionRegistry, expertAdvisorClient);
+        connectionRegistry);
     });
   }