+++ changelog.md
@@ -16,6 +16,7 @@
   - limit max concurrent synchronizations based on the number of subscribed accounts
   - implement proper rounding for position profits and account equity
   - breaking change: refactored specifications updated events
+  - breaking change: removed expert advisor API
 
 12.3.1
   - added feature to unsubscribe from market data (remove symbol from market watch)